[agent] pass the most recent revision through atomic value
Before this commit, the watcher passes the revision to the updater
through a buffered channel.  If the channel overflows, the updater
sees an old revision.  As a result, the most up-to-date changes
would not be reflected soon in some cases.

This commit changes how to pass the database revision.
The revision is now stored/load atomically through a variable.
The watcher stores the latest revision to the variable and just
notify the updater via channel.